requirements.txt
tensorflow==2.14.*
numpy==1.26.*
pandas==1.5.*
numexpr==2.8.* # pandas performance optimization
bottleneck==1.3.* # pandas performance optimization

jupyter
jupyterlab
keras_tuner==1.1.*

Flask
requests

matplotlib==3.6.*
seaborn==0.12.*
scipy
